+++ pokerole_irl/pokerole_irl/urls.py
@@ -22,4 +22,5 @@ urlpatterns = [
     path('', include('pokerole_app.urls')),
     path('accounts/', include('user_app.urls')),
     path('', include('django.contrib.auth.urls')),
+    path('worlds/', include('worlds_app.urls')),
 ] + staticfiles_urlpatterns()